requirements.txt
aiohappyeyeballs==2.3.4
aiohttp==3.10.0
aiosignal==1.3.1
anomalib==1.2.0
ansicon==1.89.0
antlr4-python3-runtime==4.9.3
anyio==4.4.0
arrow==1.3.0
async-timeout==4.0.3
attrs==23.2.0
beautifulsoup4==4.12.3
blessed==1.20.0
boto3==1.34.152
botocore==1.34.152
certifi==2024.7.4
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
contourpy==1.2.1
croniter==1.3.15
cycler==0.12.1
dateutils==0.6.12
deepdiff==7.0.1
docstring_parser==0.16
editor==1.6.6
einops==0.8.0
exceptiongroup==1.2.2
fastapi==0.88.0
filelock==3.15.4
fonttools==4.53.1
FrEIA==0.2
frozenlist==1.4.1
fsspec==2023.12.2
ftfy==6.2.0
h11==0.14.0
huggingface-hub==0.24.5
idna==3.7
imageio==2.34.2
imgaug==0.4.0
importlib_resources==6.4.0
inquirer==3.3.0
intel-openmp==2021.4.0
itsdangerous==2.2.0
Jinja2==3.1.4
jinxed==1.3.0
jmespath==1.0.1
joblib==1.4.2
jsonargparse==4.32.0
kiwisolver==1.4.5
kornia==0.7.3
kornia_rs==0.1.5
lazy_loader==0.4
lightning==2.3.3
lightning-cloud==0.5.70
lightning-utilities==0.11.6
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib
mdurl==0.1.2
mkl==2021.4.0
mpmath==1.3.0
multidict==6.0.5
mvtec-halcon==22050.0.0
networkx==3.3
numpy==1.26.3
omegaconf==2.3.0
open_clip_torch==2.26.1
opencv-python==4.5.5.62
opencv-python-headless==4.10.0.82
ordered-set==4.1.0
packaging==22.0
pandas==2.2.2
pillow==10.2.0
protobuf==5.27.3
psutil==6.0.0
pydantic==1.10.17
Pygments==2.18.0
PyJWT==2.9.0
pymodbus==3.6.8
pyparsing==3.1.2
PyQt5==5.15.11
PyQt5-Qt5==5.15.2
PyQt5_sip==12.15.0
pyserial==3.5
python-dateutil==2.9.0.post0
python-multipart==0.0.9
pytorch-lightning==2.3.3
pytz==2024.1
PyYAML==6.0.1
readchar==4.1.0
regex==2024.7.24
requests==2.32.3
rich==13.7.1
rich-argparse==1.5.2
runs==1.2.2
s3transfer==0.10.2
safetensors==0.4.3
scikit-image==0.24.0
scikit-learn==1.5.1
scipy==1.14.0
shapely==2.0.5
six==1.16.0
sniffio==1.3.1
soupsieve==2.5
starlette==0.22.0
starsessions==1.3.0
sympy==1.13.1
threadpoolctl==3.5.0
tifffile==2024.7.24
timm==1.0.7
tqdm==4.66.4
traitlets==5.14.3
types-python-dateutil==2.9.0.20240316
typeshed_client==2.7.0
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
uvicorn==0.30.4
wcwidth==0.2.13
websocket-client==1.8.0
websockets==11.0.3
xmod==1.8.1
yarl==1.9.4
